Move KeyAccessLogger
