+++ tests/test_hexary_trie.py
@@ -19,6 +19,9 @@ from trie.exceptions import (
     MissingTrieNode,
     ValidationError,
 )
+from trie.utils.db import (
+    KeyAccessLogger,
+)
 from trie.utils.nodes import (
     decode_node,
 )
@@ -151,20 +154,6 @@ def test_trie_using_fixtures(name, updates, expected, deleted, final_root):
     for invalid_proof_key in deleted:
         with pytest.raises(KeyError):
             trie.get_proof(invalid_proof_key)
-
-
-class KeyAccessLogger(dict):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.read_keys = set()
-
-    def __getitem__(self, key):
-        result = super().__getitem__(key)
-        self.read_keys.add(key)
-        return result
-
-    def unread_keys(self):
-        return self.keys() - self.read_keys
 
 
 def test_hexary_trie_saves_each_root():
